RAC-378: handle reload job execution
